plotData update
The major part of this change is to add the plotData function to proPlot along with other changes that ensure this works as desired. The purpose of this function is to change proPlot instances so that they do not plot every time a change is made by default. This is desirable for speed when running scripts. This also means that proPlot instances can be made and altered without plotting them before passing them to a conFigure instance where they will be plotted. This behaviour is controlled by the proPlot property 'InstaPlot' which is false by default.